requirements.txt
# this is the requirements.txt file...
#
# DEVELOPMENT / PRODUCTION
#
csv
python-dotenv
pgeocode
pandas
requests
 sendgrid==6.6.0

 Flask 

 gunicorn 
 #
 # TEST
 #
pytest